Chore: reorganizing files into folders.
