+++ app/src/main/java/com/sriky/popflix/customViews/AutofitRecyclerView.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-package com.sriky.popflix;
+package com.sriky.popflix.customViews;
 
 import android.content.Context;
 import android.content.res.TypedArray;
